bench(route): the verb hypothesis is refuted too, and the joint finding is the point
H2 ("where is X defined" vs "everywhere X appears"), predicate declared before
labels, run once: -2.9 pts on emitted tasks, +1.2 on prompts at 5% coverage.
On emitted tasks the sign is backwards — the definition bucket is 1
search_symbols / 4 search_text (n=5, directional). A request that names what
it wants descriptively gives a symbol-name index nothing to match.

The joint finding is worth more than either refutation: both H1 and H2 fail on
COVERAGE, not purity. Two independent properties of the query text, each
absent from 85-95% of real requests, residue at ~50% purity in both. The
information needed to route these requests is not in the query string.

Script and artifact renamed — the file holds two hypotheses now and
"shape_separation" was already the kind of stale label this session has spent
the day fixing.

No CHANGELOG entry: benchmark and roadmap only; the .289 hold is untouched.
