--- benchmarks/route_recall/shape_separation_results.json
@@ -1,75 +1,0 @@
-{
- "hypothesis": "identifier-shape predicts search_symbols over search_text",
- "verdict": "REFUTED",
- "corpus": {
-  "source": "https://raw.githubusercontent.com/rknighton/jcm-route-benchmark-corpus/main/data/route_candidates_v0.1.csv",
-  "sha256": "25d09f6ae9e8e668d1a3b9a30755cf8dbd7d063ce08868529cc152fdfe84b86e"
- },
- "emitted_tasks": {
-  "n": 35,
-  "table": {
-   "shaped": {
-    "search_symbols": 4,
-    "search_text": 1
-   },
-   "unshaped": {
-    "search_symbols": 13,
-    "search_text": 17
-   }
-  },
-  "rule_accuracy_pct": 60.0,
-  "majority_floor_pct": 51.4,
-  "lift_over_floor_pts": 8.6,
-  "coverage_pct": 14.3,
-  "residue_purity_pct": 56.7,
-  "null_p_of_shaped_bucket": 0.17
- },
- "raw_prompts": {
-  "n": 164,
-  "table": {
-   "shaped": {
-    "search_symbols": 11,
-    "search_text": 13
-   },
-   "unshaped": {
-    "search_symbols": 71,
-    "search_text": 69
-   }
-  },
-  "rule_accuracy_pct": 48.8,
-  "majority_floor_pct": 50.0,
-  "lift_over_floor_pts": -1.2,
-  "coverage_pct": 14.6,
-  "residue_purity_pct": 50.7
- },
- "per_pattern_exploratory": {
-  "snake_case": {
-   "fires": 6,
-   "search_symbols": 3
-  },
-  "camelCase": {
-   "fires": 3,
-   "search_symbols": 1
-  },
-  "PascalCase": {
-   "fires": 13,
-   "search_symbols": 7
-  },
-  "CONSTANT_CASE": {
-   "fires": 1,
-   "search_symbols": 0
-  },
-  "call_form": {
-   "fires": 1,
-   "search_symbols": 1
-  },
-  "dotted_path": {
-   "fires": 3,
-   "search_symbols": 1
-  },
-  "backticked": {
-   "fires": 0
-  }
- },
- "per_pattern_note": "Seven patterns tested on one sample. Expect one to look good by chance; none is a finding on its own."
-}
